ekf2: Update default time delay parameters
Reducing the delay parameters takes advantage of the faster processing of current generation GPS units. Testing with UBlox 7 and 8 series units in common use indicate the actual delay is between 100 and 120 sec.
Airspeed delay is very installation dependant, but the EKF is insensitive to time offsets in airspeed, so the delay has been reduced to a value that will not unnecessarily increase the size of the data buffers.
The EV delay is also highly installation dependant and we have not standard hardware on which to base setting a default parameter. For this reason it has also been set to a value that will not unnecessarily increase the buffer length.
These change result in shorter ecl EKF data buffers which reduces RAM usage and also reduces output predictor errors.